fix: etcd 3.5.7 compatibility (do not depend on /bin/sh)
Some containers where started with or had liveness checks with /bin/sh.
This fails starting with etcd 3.5.7 as it does not come with any shell.

As a consequence I also had to raise the etcd compatibility to at least
3.3 as this is the first version which has ETCDCTL_API=3 as default.

This may be solved in a different way, but 3.2 should not be used anymore.